Sizeof test

+++ sdk/Constants.go
@@ -97,7 +97,7 @@ const (
 	TCP_END_OF_BUFFER_CHAR_3                             = 0x33
 	TCP_END_OF_BUFFER_CHAR_4                             = 0x44
 
-	sizeofUINT32 = unsafe.Sizeof(0)
+	sizeofUINT32 = unsafe.Sizeof(uint32(0))
 	//UsermodeBufferSize                                        = sizeof(UINT32) + PacketChunkSize + 1
 	UsermodeBufferSize             = sizeofUINT32 + PacketChunkSize + 1
 	VERSION_MAJOR                  = 0
